docs(621): Add more repo information
This PR adds information about multiple repositories that are found in
the screwdriver-cd org. This additional info will hopefully help people
unfamiliar with the Screwdriver architecture and codebase understand how
different repos work together and where to make changes if they are so
inclined.

In addition, this change removes the npm badges that were associated
with many repos as the formatting on
http://docs.screwdriver.cd/about/contributing/ was very spaced out with
the badges present, and the information in the badges can be easily
found by clicking on the Github link for each repo.

This PR also contains some formatting and grammar modifications.