add step to run htmlproof
After using a different mechanism to find broken links the other day, I
looked for something to be run in coordination with Jekyll, and I found
the html-proof gem.

Most of the options I have included will need to remain. The
--empty-alt-ignore can go away with some alt properties in some images,
and the --file-ignore for all of the ja files can go away with some fixes
in that subdirectory.